+++ packages/Articles/src/site/components/com_articles/controllers/toolbars/article.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * Article Toolbar.
+ *
+ * @category   Anahita
+ *
+ * @author     Rastin Mehr <rastin@anahitapolis.com>
+ * @license    GNU GPLv3 <http://www.gnu.org/licenses/gpl-3.0.html>
+ *
+ * @link       http://www.GetAnahita.com
+ */
+class ComArticlesControllerToolbarArticle extends ComMediumControllerToolbarDefault
+{
+    /**
+     * Add Admin Commands for an entity.
+     */
+    public function addAdministrationCommands()
+    {
+        $this->addCommand('pin');
+        parent::addAdministrationCommands();
+    }
+}